upgrade-test Move classes of com.liferay.portal.upgrade.v7_1_x package to module
