+++ modules/apps/portal/portal-upgrade-test/src/testIntegration/java/com/liferay/portal/upgrade/v7_1_x/test/UpgradeDB2Test.java
@@ -12,23 +12,27 @@
  * details.
  */
 
-package com.liferay.portal.upgrade.v7_1_x;
+package com.liferay.portal.upgrade.v7_1_x.test;
 
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.kernel.dao.db.DB;
 import com.liferay.portal.kernel.dao.db.DBManagerUtil;
 import com.liferay.portal.kernel.dao.db.DBType;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.upgrade.BaseUpgradeDBColumnSizeTestCase;
+import com.liferay.portal.upgrade.test.BaseUpgradeDBColumnSizeTestCase;
+import com.liferay.portal.upgrade.v7_1_x.UpgradeDB2;
 
 import org.junit.Assume;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Rule;
+import org.junit.runner.RunWith;
 
 /**
  * @author Preston Crary
  */
+@RunWith(Arquillian.class)
 public class UpgradeDB2Test extends BaseUpgradeDBColumnSizeTestCase {
 
 	@ClassRule
